Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
torch==2.3.1
torchvision==0.18.1
tensorflow-cpu==2.15.0
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.7.1
email-validator==2.1.1
